Move Embassy server config to helper file
